feat(stellar): wallet creation, XLM transfers, and tx verification
