+++ package.json
@@ -41,6 +41,7 @@
     "@nestjs/throttler": "^6.5.0",
     "@nestjs/typeorm": "^11.0.0",
     "@nestjs/websockets": "^11.1.17",
+    "@stellar/stellar-sdk": "^16.0.1",
     "archiver": "^7.0.1",
     "axios": "^1.7.9",
     "bcrypt": "^5.1.1",
