+++ src/blockchain/stellar/stellar.module.ts
@@ -1,8 +1,1 @@
-import { Module } from '@nestjs/common';
-import { StellarService } from './stellar.service';
-
-@Module({
-  providers: [StellarService],
-  exports: [StellarService],
-})
-export class StellarModule {}
+export { StellarModule } from '../../modules/stellar/stellar.module';
