+++ src/blockchain/stellar/stellar.types.ts
@@ -1,27 +1,1 @@
-import { Transaction } from 'stellar-sdk';
-
-export interface GenerateWalletResult {
-  publicKey: string;
-  secretKey: string;
-}
-
-export interface CreateTransactionParams {
-  sourcePublicKey: string;
-  operations: any[];
-  memo?: string;
-  memoType?: string;
-  userId?: string; // Optional for audit logging
-}
-
-export interface VerifyTransactionResult {
-  status: 'PENDING' | 'SUCCESS' | 'FAILED';
-  details?: any;
-}
-
-export interface WalletBalanceResult {
-  asset: string;
-  balance: string;
-  assetIssuer?: string;
-}
-
-export type StellarTransaction = Transaction;
+export * from '../../modules/stellar/stellar.types';
